adding controller, refactor

+++ app/repository.py
@@ -1,20 +1,34 @@
-from fastapi import HTTPException
+from sqlalchemy.exc import SQLAlchemyError
+from app.database import Base, engine, SessionLocal
+
+# creates database
+Base.metadata.create_all(engine)
+
+
+def get_session():
+    session = SessionLocal()
+    try:
+        yield session
+    finally:
+        session.close()
 
 
 class SQLiteRepository:
     @staticmethod
-    def get_all(model, session):
-        docs = session.query(model).all()
-        return docs
+    def fetch_all(model, session):
+        try:
+            data = session.query(model).all()
+            return data
+        except SQLAlchemyError:
+            return Exception
 
     @staticmethod
-    def get_one(object_id, session, model):
-        object = session.query(model).get(object_id)
-        if not object:
-            raise HTTPException(
-                status_code=404, detail=f"Object with ID {object_id} not found"
-            )
-        return object
+    def fetch_one(object_id, session, model):
+        data = session.query(model).get(object_id)
+        if not data:
+            return Exception
+        else:
+            return data
 
     @staticmethod
     def add_object(session, object):
@@ -24,8 +38,9 @@ class SQLiteRepository:
 
     @staticmethod
     def delete_object(object_id, model, session):
-        object = SQLiteRepository.get_one(object_id, session, model)
+        object = SQLiteRepository.fetch_one(object_id, session, model)
+        if object == Exception:
+            return object
         session.delete(object)
         session.commit()
         session.close()
-        return "Object was deleted!"
